+++ sdv/utils/__init__.py
@@ -1,5 +1,5 @@
 """Utils module."""
 
-from sdv.utils.utils import drop_unknown_references
+from sdv.utils.utils import drop_unknown_references, get_random_sequence_subset
 
-__all__ = ('drop_unknown_references',)
+__all__ = ('drop_unknown_references', 'get_random_sequence_subset')
