+++ sdv/utils/utils.py
@@ -3,6 +3,7 @@
 import sys
 from copy import deepcopy
 
+import numpy as np
 import pandas as pd
 
 from sdv._utils import _validate_foreign_keys_not_null
@@ -60,3 +61,71 @@ def drop_unknown_references(data, metadata, drop_missing_values=True, verbose=Tr
             sys.stdout.write('\n'.join([success_message, '', summary_table.to_string(index=False)]))
 
         return result
+
+
+def get_random_sequence_subset(
+    data,
+    metadata,
+    num_sequences,
+    max_sequence_length=None,
+    long_sequence_subsampling_method='first_rows',
+):
+    """Subsample sequential data based on a number of sequences.
+
+    Args:
+        data (pandas.DataFrame):
+            The sequential data.
+        metadata (SingleTableMetadata):
+            A SingleTableMetadata object describing the data.
+        num_sequences (int):
+            The number of sequences to subsample.
+        max_sequence_length (int):
+            The maximum length each subsampled sequence is allowed to be. Defaults to None. If
+            None, do not enforce any max length, meaning that entire sequences will be sampled.
+            If provided all subsampled sequences must be <= the provided length.
+        long_sequence_subsampling_method (str):
+            The method to use when a selected sequence is too long. Options are:
+            - (default) first_rows: Keep the first n rows of the sequence, where n is the max
+            sequence length.
+            - last_rows: Keep the last n rows of the sequence, where n is the max sequence length.
+            - random: Randomly choose n rows to keep within the sequence. It is important to keep
+            the randomly chosen rows in the same order as they appear in the original data.
+    """
+    if long_sequence_subsampling_method not in ['first_rows', 'last_rows', 'random']:
+        raise ValueError(
+            'long_sequence_subsampling_method must be one of "first_rows", "last_rows" or "random"'
+        )
+
+    sequence_key = metadata.sequence_key
+    if not sequence_key:
+        raise ValueError(
+            'Your metadata does not include a sequence key. A sequence key must be provided to '
+            'subset the sequential data.'
+        )
+
+    if sequence_key not in data:
+        raise ValueError(
+            'Your provided sequence key is not in the data. This is required to get a subset.'
+        )
+
+    selected_sequences = np.random.permutation(data[sequence_key])[:num_sequences]
+    subset = data[data[sequence_key].isin(selected_sequences)].reset_index(drop=True)
+    if max_sequence_length:
+        grouped_sequences = subset.groupby(sequence_key)
+        if long_sequence_subsampling_method == 'first_rows':
+            return grouped_sequences.head(max_sequence_length).reset_index(drop=True)
+        elif long_sequence_subsampling_method == 'last_rows':
+            return grouped_sequences.tail(max_sequence_length).reset_index(drop=True)
+        else:
+            subsetted_sequences = []
+            for _, group in grouped_sequences:
+                if len(group) > max_sequence_length:
+                    idx = np.random.permutation(len(group))[:max_sequence_length]
+                    idx.sort()
+                    subsetted_sequences.append(group.iloc[idx])
+                else:
+                    subsetted_sequences.append(group)
+
+            return pd.concat(subsetted_sequences, ignore_index=True)
+
+    return subset
