+++ tests/integration/utils/test_utils.py
@@ -5,9 +5,10 @@ import numpy as np
 import pandas as pd
 import pytest
 
+from sdv.datasets.demo import download_demo
 from sdv.errors import InvalidDataError
 from sdv.metadata import MultiTableMetadata
-from sdv.utils import drop_unknown_references
+from sdv.utils import drop_unknown_references, get_random_sequence_subset
 
 
 @pytest.fixture
@@ -140,3 +141,54 @@ def test_drop_unknown_references_not_drop_missing_values(metadata, data):
     pd.testing.assert_frame_equal(cleaned_data['child'], data['child'].iloc[:4])
     assert pd.isna(cleaned_data['child']['parent_id']).any()
     assert len(cleaned_data['child']) == 4
+
+
+def test_get_random_sequence_subset():
+    """Test that the sequences are subsetted and properly clipped."""
+    # Setup
+    data, metadata = download_demo(modality='sequential', dataset_name='nasdaq100_2019')
+
+    # Run
+    subset = get_random_sequence_subset(data, metadata, num_sequences=3, max_sequence_length=5)
+
+    # Assert
+    selected_sequences = subset[metadata.sequence_key].unique()
+    assert len(selected_sequences) == 3
+    for sequence_key in selected_sequences:
+        pd.testing.assert_frame_equal(
+            subset[subset[metadata.sequence_key] == sequence_key].reset_index(drop=True),
+            data[data[metadata.sequence_key] == sequence_key].head(5).reset_index(drop=True),
+        )
+
+
+def test_get_random_sequence_subset_random_clipping():
+    """Test that the sequences are subsetted and properly clipped.
+
+    If the long_sequence_sampling_method is set to 'random', the selected sequences should be
+    subsampled randomly, but maintain the same order.
+    """
+    # Setup
+    data, metadata = download_demo(modality='sequential', dataset_name='nasdaq100_2019')
+
+    # Run
+    subset = get_random_sequence_subset(
+        data,
+        metadata,
+        num_sequences=3,
+        max_sequence_length=5,
+        long_sequence_subsampling_method='random',
+    )
+
+    # Assert
+    selected_sequences = subset[metadata.sequence_key].unique()
+    assert len(selected_sequences) == 3
+    for sequence_key in selected_sequences:
+        selected_sequence = subset[subset[metadata.sequence_key] == sequence_key]
+        assert len(selected_sequence) <= 5
+        subset_data = data[
+            data['Date'].isin(selected_sequence['Date'])
+            & data['Symbol'].isin(selected_sequence['Symbol'])
+        ]
+        pd.testing.assert_frame_equal(
+            subset_data.reset_index(drop=True), selected_sequence.reset_index(drop=True)
+        )
